requirements.txt
flask==2.0.1
monai==1.2.0
torch==2.0.0
pillow==9.5.0
python-dotenv==0.19.0
openai==1.0.0
numpy==1.23.5
pydicom==2.3.1 